Commercial features - phase 2
