+++ esbuild.config.js
@@ -72,6 +72,11 @@ esbuild
     },
     inject: ["./esbuild-shim.js"],
     outfile: "ssr.js",
+    alias: {
+      "@fider": "./public",
+      "@commercial": "./commercial",
+      "@locale": "./locale",
+    },
     plugins: [emptyCSS, emptySVG, babelPlugin()],
   })
   .catch(() => process.exit(1))
