+++ public/components/index.tsx
@@ -8,7 +8,7 @@ export * from "./SignInModal"
 export * from "./RSSModal"
 export * from "./VoteCounter"
 export * from "./NotificationIndicator"
-export * from "./ModerationIndicator"
+export * from "@commercial/components/ModerationIndicator"
 export * from "./UserMenu"
 export * from "./Reactions"
 export * from "./ReadOnlyNotice"
